+++ src/entities/registration-package/index.ts
@@ -1,0 +1,6 @@
+export { registrationPackagesQueryOptions } from '#/entities/registration-package/api/packages'
+export type {
+  ActivePackage,
+  PackageList,
+  RegistrationPackage,
+} from '#/entities/registration-package/model/types'