+++ src/features/activate-package/index.ts
@@ -1,0 +1,1 @@
+export { useActivatePackage } from '#/features/activate-package/model/use-activate'